+++ src/test/java/technology/transceptor/QPackTest.java
@@ -1,4 +1,4 @@
-package transceptor.technology;
+package technology.transceptor;
 
 import java.lang.reflect.Array;
 import java.util.ArrayList;
@@ -13,7 +13,7 @@ import java.util.Queue;
 import java.util.Stack;
 import org.junit.Test;
 import static org.junit.Assert.*;
-import static transceptor.technology.Types.*;
+import static technology.transceptor.Types.*;
 
 /**
  *
